+++ routes/web.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\OCRController;
 use App\Http\Controllers\User\LoginController;
 use App\Http\Controllers\User\PostController;
 use App\Http\Controllers\User\UserController;
+use App\Http\Controllers\CommentController;
 
 Route::get('/', [MainController::class, 'index'])->name('threads.index');
 Route::post('/post', [MainController::class, 'post'])->name('threads.post');
@@ -26,17 +27,23 @@ Route::get('/login', [LoginController::class, 'showLoginForm'])->name('login');
 Route::post('/login', [LoginController::class, 'login']);
 Route::post('/logout', [LoginController::class, 'logout'])->name('logout');
 
-Route::get('/', [PostController::class, 'home'])->middleware('auth')->name('home')->middleware('auth.custom');;
+Route::get('/', [PostController::class, 'home'])->middleware('auth')->name('home')->middleware('auth.custom');
+Route::get('/posts/create', [PostController::class, 'create'])->name('posts.create');
+Route::post('/posts', [PostController::class, 'store'])->name('posts.store');
 
-Route::resource('posts', PostController::class);
+// Profile user (direct mention)
+Route::get('/{user:mention}', [UserController::class, 'show'])->name('users.show');
+Route::get('/{user:mention}/edit', [UserController::class, 'edit'])->name('users.edit')->middleware('auth');
+Route::put('/{user:mention}', [UserController::class, 'update'])->name('users.update')->middleware('auth');
+Route::delete('/{user:mention}', [UserController::class, 'destroy'])->name('users.destroy')->middleware('auth');
 
-Route::prefix('users')->group(function () {
-    Route::get('/', [UserController::class, 'index'])->name('users.index');       // danh sách
-    Route::get('/create', [UserController::class, 'create'])->name('users.create'); // form tạo
-    Route::post('/', [UserController::class, 'store'])->name('users.store');       // lưu mới
-
-    Route::get('/{mention}', [UserController::class, 'show'])->name('users.show'); // xem profile
-    Route::get('/{mention}/edit', [UserController::class, 'edit'])->name('users.edit'); // form sửa
-    Route::put('/{mention}', [UserController::class, 'update'])->name('users.update'); // cập nhật
-    Route::delete('/{mention}', [UserController::class, 'destroy'])->name('users.destroy'); // xoá
+// Post của user
+Route::prefix('{user:mention}/{post:slug}')->group(function () {
+    Route::get('/', [PostController::class, 'show'])->name('posts.show');
+    Route::get('/edit', [PostController::class, 'edit'])->name('posts.edit')->middleware('auth');
+    Route::put('/', [PostController::class, 'update'])->name('posts.update')->middleware('auth');
+    Route::delete('/', [PostController::class, 'destroy'])->name('posts.destroy')->middleware('auth');
+    // Comment routes
+    Route::post('/comments', [CommentController::class, 'store'])->name('comments.store');
+    Route::delete('/comments/{comment}', [CommentController::class, 'destroy'])->name('comments.destroy');
 });
